Create Create object with a factory function (generic).js
Create object with a factory function (generic).js uses an object prototype in the Object.create() function to create new objects with the prototype as the pattern.
We also have a showObject() function to encapsulate a readable list output of the properties for this example object -- which is easily extended to more complex objects. 
The showObject is used to easily confirm property changes in an object after creating the object with the makeObject factory function.
This example is the generic version of that found in Chapter 2 of the book "Single Page Web Applications".